requirements.txt
absl-py==1.0.0
affine==2.3.0
alabaster==0.7.8
anyio==3.4.0
anykeystore==0.2
apex==0.9.10.dev0
appdirs==1.4.3
apturl==0.5.2
argon2-cffi==21.1.0
asciitree==0.3.3
asgiref==3.4.1
attrs==21.2.0
autopep8==1.6.0
Babel==2.6.0
backcall==0.2.0
backports.entry-points-selectable==1.1.1
bcrypt==3.1.7
beautifulsoup4==4.10.0
bleach==4.1.0
blinker==1.4
bqplot==0.12.32
branca==0.4.2
Brlapi==0.7.0
cachetools==4.2.4
certifi==2019.11.28
cffi==1.15.0
cftime==1.5.1.1
chardet==3.0.4
Click==7.0
click-plugins==1.1.1
cligj==0.7.2
cloudpickle==2.0.0
colorama==0.4.3
colour==0.1.5
command-not-found==0.3
cryptacular==1.6.2
cryptography==2.8
cupshelpers==1.0
cycler==0.11.0
dask==2021.11.2
DatastreamDSWS==1.0.10
DateTime==4.3
dbus-python==1.2.16
debugpy==1.5.1
decorator==5.1.0
defer==1.0.6
defusedxml==0.7.1
distlib==0.3.3
distro==1.4.0
distro-info===0.23ubuntu1
docopt==0.6.2
docutils==0.16
duplicity==0.8.12.0
earthengine-api==0.1.290
earthpy==0.9.4
entrypoints==0.3
fastapi==0.70.0
fasteners==0.14.1
feedparser==5.2.1
ffmpeg-python==0.2.0
filelock==3.3.2
Fiona==1.8.20
folium==0.12.1.post1
fonttools==4.28.2
fsspec==2021.11.1
future==0.18.2
geeadd==0.5.5
geemap==0.11.0
geocoder==1.38.1
geojson==2.5.0
geopandas==0.10.2
gitsome==0.8.0
google-api-core==1.31.4
google-api-python-client==1.12.8
google-auth==2.3.3
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
google-cloud-core==2.2.1
google-cloud-storage==1.43.0
google-crc32c==1.3.0
google-resumable-media==2.1.0
googleapis-common-protos==1.53.0
googledrivedownloader==0.4
greenlet==1.1.2
grpcio==1.42.0
h11==0.12.0
httplib2==0.14.0
httplib2shim==0.0.3
httptools==0.2.0
hupper==1.10.3
idna==2.8
imagecodecs==2021.11.20
imageio==2.12.0
imagesize==1.2.0
importlib-metadata==4.8.2
importlib-resources==5.4.0
intake==0.6.4
intake-xarray==0.5.0
ipyevents==2.0.1
ipyfilechooser==0.6.0
ipykernel==6.5.1
ipyleaflet==0.15.0
ipynb-py-convert==0.4.6
ipython==7.29.0
ipython-genutils==0.2.0
ipytree==0.2.1
ipywidgets==7.6.5
jedi==0.18.1
Jinja2==2.10.1
joblib==1.1.0
json5==0.9.6
jsonschema==4.2.1
jupyter==1.0.0
jupyter-client==7.0.6
jupyter-console==6.4.0
jupyter-core==4.9.1
jupyter-server==1.13.3
jupyterlab==3.2.8
jupyterlab-pygments==0.1.2
jupyterlab-server==2.10.3
jupyterlab-widgets==1.0.2
keyring==18.0.1
kiwisolver==1.3.2
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
locket==0.2.1
lockfile==0.12.2
logzero==1.7.0
louis==3.12.0
macaroonbakery==1.3.1
Mako==1.1.0
Markdown==3.3.6
MarkupSafe==1.1.0
matplotlib==3.5.0
matplotlib-inline==0.1.3
mistune==0.8.4
monotonic==1.5
more-itertools==4.2.0
msgpack==1.0.3
mss==6.1.0
munch==2.5.0
nbclassic==0.3.5
nbclient==0.5.9
nbconvert==6.3.0
nbformat==5.1.3
nest-asyncio==1.5.1
netCDF4==1.5.8
netifaces==0.10.4
networkx==2.6.3
notebook==6.4.6
numcodecs==0.9.1
numpy==1.21.4
numpydoc==0.7.0
oauthlib==3.1.0
olefile==0.46
opencv-python==4.5.4.60
OWSLib==0.25.0
packaging==20.3
palettable==3.3.0
pandas==1.3.4
pandocfilters==1.5.0
paramiko==2.6.0
parso==0.8.2
partd==1.2.0
PasteDeploy==2.1.1
pbkdf2==1.3
pexpect==4.6.0
pickleshare==0.7.5
Pillow==7.0.0
pipenv==11.9.0
plaster==1.0
plaster-pastedeploy==0.7
platformdirs==2.4.0
plotly==5.5.0
ply==3.11
prometheus-client==0.12.0
prompt-toolkit==2.0.10
protobuf==3.6.1
psutil==5.8.0
psycopg2==2.9.3
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.16.2
pycodestyle==2.8.0
pycparser==2.21
PyCRS==1.0.2
pycups==1.9.73
pydantic==1.8.2
Pygments==2.10.0
PyGObject==3.36.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pynvml==11.4.1
pyparsing==2.4.6
pyproj==3.3.0
pyramid==2.0
pyramid-mailer==0.15.1
pyRFC3339==1.1
pyrsistent==0.18.0
pyshp==2.1.3
python-apt==2.0.0+ubuntu0.20.4.6
python-box==5.4.1
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
python-dotenv==0.19.2
python-multipart==0.0.5
python3-openid==3.2.0
pytorch-metric-learning==1.1.0
pytz==2019.3
PyWavelets==1.2.0
pyxdg==0.26
PyYAML==5.3.1
pyzmq==22.3.0
qtconsole==5.2.0
QtPy==1.11.2
rasterio==1.2.10
ratelim==0.1.6
reportlab==3.5.34
repoze.sendmail==4.4.1
requests==2.22.0
requests-oauthlib==1.3.0
requests-unixsocket==0.2.0
rioxarray==0.9.0
roman==2.0.0
rsa==4.8
Rtree==0.9.7
ruamel.yaml==0.17.17
ruamel.yaml.clib==0.2.6
sankee==0.0.7
scikit-image==0.18.3
scikit-learn==1.0.2
scipy==1.7.3
screen-resolution-extra==0.0.0
SecretStorage==2.3.1
Send2Trash==1.8.0
setproctitle==1.1.10
setuptools-scm==6.3.2
Shapely==1.8.0
simplejson==3.16.0
six==1.14.0
sklearn==0.0
sniffio==1.2.0
snuggs==1.4.7
soupsieve==2.3.1
Sphinx==1.8.5
SQLAlchemy==1.4.28
starlette==0.16.0
systemd-python==234
tenacity==8.0.1
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
terminado==0.12.1
testpath==0.5.0
threadpoolctl==3.0.0
tifffile==2021.11.2
timm==0.4.12
toml==0.10.2
tomli==1.2.2
toolz==0.11.2
torch==1.10.0
torch-tb-profiler==0.3.1
torchvision==0.11.1
tornado==6.1
tqdm==4.62.3
traitlets==5.1.1
traittypes==0.2.1
transaction==3.0.1
translationstring==1.4
typing-extensions==4.0.0
ubuntu-advantage-tools==27.5
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
uritemplate==3.0.1
urllib3==1.25.8
usb-creator==0.3.7
uvicorn==0.15.0
uvloop==0.16.0
velruse==1.1.1
venusian==3.0.0
virtualenv==20.10.0
virtualenv-clone==0.5.7
voila==0.3.0
wadllib==1.3.3
watchgod==0.7
wcwidth==0.1.8
webencodings==0.5.1
WebOb==1.8.7
websocket-client==1.2.3
websockets==10.1
Werkzeug==2.0.2
whitebox==2.0.3
whiteboxgui==0.6.0
widgetsnbextension==3.5.2
WTForms==3.0.0
wtforms-recaptcha==0.3.2
xarray==0.20.1
xgboost==1.5.1
xkit==0.0.0
xonsh==0.9.13
xyzservices==2021.11.0
zarr==2.10.3
zipp==3.6.0
zope.deprecation==4.4.0
zope.interface==5.4.0
zope.sqlalchemy==1.6
